mcp_ias_auth/requirements.txt
fastmcp
fastapi
uvicorn
requests
pyjwt
cryptography
